pizzeria_backend/requirements.txt
flask==2.3.2
python-telegram-bot==20.3
uvicorn==0.22.0
flask-cors==4.0.0
asgiref==3.7.2
firebase-admin==6.1.0
werkzeug==2.3.4